add to list books per idiom

+++ src/main/java/com/socamaru/literalura/repository/models/AuthorDTO.java
@@ -2,8 +2,8 @@ package com.socamaru.literalura.repository.models;
 
 public record AuthorDTO(
 	String name,
-	int birthYear,
-	int deathYear
+	Long birthYear,
+	Long deathYear
 ) {
